+++ CHANGELOG.rst
@@ -1,10 +1,20 @@
 CHANGELOG
 =========
 
-6.0.0a7 (unreleased)
+6.0.0a8 (unreleased)
 --------------------
 
 - Fix command 'create'
+  [masipcat]
+
+
+6.0.0a7 (2020-01-17)
+--------------------
+
+- Better error handling on redis connection issues
+  [vangheem]
+
+- Run _update_from_pytest_markers() after configuring db settings
   [masipcat]
 
 - Fix validating array params in query parameters [lferran]
@@ -21,8 +31,10 @@ CHANGELOG
 - Fix automatic type conversion on nested fields. Fixes #832
   [vangheem]
 
+
 6.0.0a6 (2020-01-13)
 --------------------
+
 - Fix bug on swagger with endpoints without explicit security declarations
   [jordic]